+++ example/lib/map_with_separate_popups.dart
@@ -26,15 +26,15 @@ class _MapWithSeparatePopupsState extends State<MapWithSeparatePopups> {
   void initState() {
     super.initState();
     _markersA = [
-      LatLng(44.421, 10.404),
-      LatLng(45.683, 10.839),
-      LatLng(45.246, 5.783),
+      const LatLng(44.421, 10.404),
+      const LatLng(45.683, 10.839),
+      const LatLng(45.246, 5.783),
     ].map((latLng) => _markerFrom(latLng, Colors.black)).toList();
 
     _markersB = [
-      LatLng(39.421, 10.404),
-      LatLng(40.683, 10.839),
-      LatLng(40.246, 5.783),
+      const LatLng(39.421, 10.404),
+      const LatLng(40.683, 10.839),
+      const LatLng(40.246, 5.783),
     ].map((latLng) => _markerFrom(latLng, Colors.blue)).toList();
   }
 
@@ -69,7 +69,7 @@ class _MapWithSeparatePopupsState extends State<MapWithSeparatePopups> {
         child: FlutterMap(
           options: MapOptions(
             zoom: 5.0,
-            center: LatLng(44.421, 10.404),
+            center: const LatLng(44.421, 10.404),
             onTap: (_, __) => _popupLayerController
                 .hideAllPopups(), // Hide popup when the map is tapped.
           ),